Fixed code structure
