+++ alphazero/__init__.py
@@ -2,6 +2,6 @@
 
 """AlphaZero implementation for various games."""
 from ._game import *  # noqa
-from ._utils import *  # noqa
 from .model import *  # noqa
 from ._boardconverters import *  # noqa
+from .storage import *  # noqa
