+++ alphazero/training/__init__.py
@@ -1,0 +1,2 @@
+from ._data import * # noqa
+from ._training import # noqa
